made coding refactoring and some styling
