+++ client/src/assets/scss/main.scss
@@ -192,3 +192,12 @@ input[type=submit] {
   color: #A9A9A9;
   opacity: 1;
 }
+.headers{
+  font-family:"Times New Roman", Times, serif;
+}
+
+.disclaimer{
+  font-family: "Lucida Console", "Courier New", monospace;
+  font-size: small;
+  font-style: italic;
+}
