+++ client/src/components/PuzzlePage.js
@@ -4,7 +4,7 @@ const PuzzlePage = (props) => {
 
   return(
     <div className="callout primary">
-      <h1>Here are some puzzles!</h1>
+      <h1 className="headers">Here are some puzzles!</h1>
       <ul>
         <li><a href="/puzzle/1">Easy Puzzle!</a></li>
         <li><a href="/puzzle/2">Another Easy Puzzle!</a></li>
@@ -17,8 +17,8 @@ const PuzzlePage = (props) => {
         <li><a href="/puzzle/9">The Last Hard Puzzle!</a></li>
       </ul>
 
-      <h1>Randomly Generated Puzzles from Sugoku!</h1> 
-      <h5>Note: Due to the inconsistency of the API, know that some of these puzzles might not actually be completeable.</h5>
+      <h1 className="headers">Randomly Generated Puzzles from Sugoku!</h1> 
+      <h5 className="disclaimer">Note: Due to the inconsistency of the API, know that some of these puzzles might not actually be completeable.</h5>
       <ul>
         <li><a href="/randomPuzzle/easy">Randomly Generated Easy Puzzle!</a></li>
         <li><a href="/randomPuzzle/medium">Randomly Generated Medium Puzzle!</a></li>
