+++ server/src/services/validatePuzzleSquares.js
@@ -1,18 +1,26 @@
-const cleanUserSubmission = formInput => {
-  for(const box in formInput){
-    let adjustedBox = formInput[box]
-    
-    if(formInput[box].length < 9){
-      for(let x = 0; x< 9; x++){
-        formInput[box][x] = 0
+const validatePuzzleSquares = formInput => {
+  Object.keys(formInput).forEach(field => {
+    for(const box in formInput){
+      let adjustedBox = formInput[box]
+      
+      if(formInput[box].length < 9){
+        for(let x = 0; x< 9; x++){
+          formInput[box][x] = 0
+        }
       }
-    }
-
-    formInput[box].forEach((square, index) => {
-        adjustedBox[index] = parseInt(square)
-        return adjustedBox
+  
+      formInput[box].forEach((square, index) => {
+          adjustedBox[index] = parseInt(square)
+          return adjustedBox
+        })
+  
+      formInput[box].forEach((square) => {
+        if( isNaN( Number(square) ) ){
+          delete formInput[box]
+        }
       })
-  }
+    }
+  })
   return formInput
 }
-export default cleanUserSubmission
+export default validatePuzzleSquares
